+++ internal/web/config.go
@@ -53,7 +53,11 @@ func configHandler(c *gin.Context) {
 		c.HTML(http.StatusOK, "header.html", guiData)
 		c.HTML(http.StatusOK, "config.html", guiData)
 	} else {
-		loginScreen(c, appConfig.Host+":"+appConfig.PortConf, "Config") // login.go
+		var targetStruct models.TargetStruct
+		targetStruct.Target = appConfig.Host + ":" + appConfig.PortConf
+		targetStruct.Name = "Config"
+
+		loginScreen(c, targetStruct) // login.go
 	}
 }