+++ internal/web/login.go
@@ -13,24 +13,29 @@ import (
 )
 
 func loginHandler(c *gin.Context) {
-	var target, name string
+	var authOk bool
 
 	proxyAddr := c.MustGet("proxyAddr").(string)
 	targetStruct, ok := targetMap[proxyAddr]
 
-	if ok {
-		target = targetStruct.Target
-		name = targetStruct.Name
+	if !ok {
+		targetStruct.Target = appConfig.Target
+		targetStruct.Name = "Default"
+
+		authOk = auth.Auth(c, &authConf)
 	} else {
-		target = appConfig.Target
-		name = "Default"
+		username, sesOk := auth.GetCurrentUser(c)
+		_, ok := targetStruct.Users[username]
+
+		if sesOk && (ok || username == authConf.User) {
+			authOk = true
+		}
 	}
 
-	authOk := auth.Auth(c, &authConf)
 	if authOk {
-		reverseProxy(c, target)
+		reverseProxy(c, targetStruct.Target)
 	} else {
-		loginScreen(c, target, name)
+		loginScreen(c, targetStruct)
 	}
 }
 
@@ -45,24 +50,28 @@ func reverseProxy(c *gin.Context, target string) {
 	proxy.ServeHTTP(c.Writer, c.Request)
 }
 
-func loginScreen(c *gin.Context, target, name string) {
+func loginScreen(c *gin.Context, targetStruct models.TargetStruct) {
 	var guiData models.GuiData
 
 	username := c.PostForm("username")
 	password := c.PostForm("password")
 
-	if username == authConf.User && auth.MatchPasswords(authConf.Password, password) {
+	currentAuth, ok := checkUsername(targetStruct, username, password)
 
-		msg := "User '" + username + "' logged in from " + c.Request.RemoteAddr + ". Session expires in " + authConf.Expire.String() + ". Target: " + target + " (" + name + ")"
+	if ok {
+
+		msg := "User '" + username + "' logged in from " + c.Request.RemoteAddr + ". Session expires in " + authConf.Expire.String() + ". Target: " + targetStruct.Target + " (" + targetStruct.Name + ")"
 		log.Println("INFO:", msg)
 		go notify.Shout("ForAuth: "+msg, appConfig.Notify)
 
-		auth.StartSession(c)
+		log.Println("REQUEST:", c.Request)
 
-		c.Redirect(http.StatusFound, "/")
+		auth.StartSession(c, currentAuth)
+
+		c.Redirect(http.StatusFound, c.Request.Referer())
 	} else {
 		if username != "" {
-			msg := "Incorrect login attempt by '" + username + "' with password '" + password + "' logged in from " + c.Request.RemoteAddr + ". Target: " + target + " (" + name + ")"
+			msg := "Incorrect login attempt by '" + username + "' with password '" + password + "' logged in from " + c.Request.RemoteAddr + ". Target: " + targetStruct.Target + " (" + targetStruct.Name + ")"
 			log.Println("WARNING:", msg)
 			go notify.Shout("ForAuth: "+msg, appConfig.Notify)
 		}
@@ -72,4 +81,21 @@ func loginScreen(c *gin.Context, target, name string) {
 		c.HTML(http.StatusOK, "header.html", guiData)
 		c.HTML(http.StatusOK, "login.html", guiData)
 	}
+}
+
+func checkUsername(targetStruct models.TargetStruct, username, password string) (auth.Conf, bool) {
+
+	if username == authConf.User && auth.MatchPasswords(authConf.Password, password) {
+		return authConf, true
+	}
+
+	targetAuth, ok := targetStruct.Users[username]
+	if ok && (username == targetAuth.User &&
+		auth.MatchPasswords(targetAuth.Password, password)) {
+
+		targetAuth.Expire = auth.ToTime(targetAuth.ExpStr)
+		return targetAuth, true
+	}
+
+	return authConf, false
 }