+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.2.1] - 2026-02-
+### Changed
+- Multiuser for targets
+- Redirect to Referer() instead of "/"
+
 ## [0.1.5] - 2025-09-12
 ### Changed
 - Upgraded to `go-1.25.1`